requirements.txt
atlassian-python-api==1.17.6
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
future==0.18.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
PyJWT==1.7.1
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
six==1.15.0
urllib3==1.25.11
webexteamssdk==1.6
Werkzeug==1.0.1
